Drop Handler, Improved Tile Breaking, Explosions
Added a drop handler to improve item drops. This allows for drops to be handled in a component that can be attached to anything.

Improved tile breaking and made it a callable function in the GameManager singleton

Added "explosions". This works by taking in a radius and creating an array of Vector2 points of where to check if a tile exists. It checks them in a circle and the points are separated by 1 tile unit, or 16px. The circle shape is a pixel art circle, so the radius is in pixels.

Added grenades to mess around with explosions.

Project working time is now at 15.5 hours.