New ResourceNode Spawning, Beginning a Inventory System
Added new art for the tiles. Thanks Neomenight!

Reworked the resource node spawning mechanic. Instead of having a collider that moves around to each resource node spawning attempt (which was very slow), it now checks the distance from the position of the new attempted node to the last spawned node and if it is greater than or equal to the range in the parameters, it passes the check and can spawn there. This is significantly faster than it previously was. It used to be at least 0.05s per check (painfully slow) since there were usually around 70 nodes spawning for a 256 tile wide area), but now it is fast and happens by the time the game even loads in.

 Finally, began work on an inventory system. Currently, there is a function for adding items to the inventory's item array, which makes sure items don't go past their max_quantity amount. It works recursively to guarantee this.

I am very happy with the progress so far. Lastly, we are now at 24.5 hours (roughly) on the project! I really like tracking the hours on a spreadsheet, it is cool to see how much progress I am getting done in the time I work on it.