Fixed quicksort, drag right click, test chest
I forgot to list it in the last commit, but I added a drag feature to right click. It is similar to Minecraft's right click drag feature, which allows you to quickly put a single item across many item slots in an inventory.

Secondly, I fixed the quicksort algorithm in the InventoryBase class. It is now a proper quicksort, so the time complexity is now average O(n log n) instead of always being O (n^2) ! Yippie efficiency!

Thirdly, in InventoryUI, I tried to make it a bit more efficient by taking out the process function and putting the input nodes in the _input fuction. Hopefully it should be a little better now. I want to see if I can make InventoryUI only process when it is going to be opened. I don't think it would be a huge lag issue but ya never know. I am still not fully sure if I will implement this idea of making it not always be processed normally.

Lastly, I added a test chest just to see how alternate inventories would work and it works great! Happy to see how well this came out.