Block breaking using tools, new sprites
Howdy ho!

In this update I've added block breaking using tools and implemented a way to store tile damages and other info about tiles. I am pretty happy with this system. It basically has a Dictionary with the position for the key and the info as the value. Damaging blocks just calls the damage function of the tile's damage tracker.

Secondly, I added some new sprites. I am liking 16x16 size sprites for the tiles. I experimented with 8x8, but found them to be too small.

Third, I am eager to work on the smithing system for the game, but there are a few little things I want to add before I get to that. First, I want to item interactions with the world. So placing tiles, breaking tiles, breaking and placing block entities such as chests and trees, etc. Once all of that is done and I feel that item usage is at a good state, I will finally work on smithing! I am so excited as that is the main point of the game and will give it some identity.

Project time is at ~70 hours. Phew! Also, I leave for study abroad in 4 days on Thursday (well, 3 cause it's past midnight.)

Anyways, goodnight, y'all!

P.S I will be switching to a private repository soon since this game has expanded past the survival-test experiment I originally made it for. I will continue my work in my first large indie title, Hammerfall! Woop woop!

Anways, thanks for reading everyone, have a good one. (Jan 15, 2024)